Remove ReservationTImerContext to simplify ReservationTimer component Move actions to update seats reservation to a separate hook
