+++ src/domain/organizations/utils.ts
@@ -2,7 +2,6 @@ import React from 'react';
 
 import addParamsToQueryString from '../../utils/addParamsToQueryString';
 import getValue from '../../utils/getValue';
-import replaceParamsToQueryString from '../../utils/replaceParamsToQueryString';
 import stripLanguageFromPath from '../../utils/stripLanguageFromPath';
 import { assertUnreachable } from '../../utils/typescript';
 import {
